// src/Models/DossierPieceJointe.ts
export interface DossierPieceJointe {
    id: number;       
    dossierPieceJointeLibelle : string;    
    dossierPieceJointeLibelleArabe : string; 
    dossierPieceJointeObligatoire: number; 
    dossierPieceJointeOriginale: number;    
    dossierPieceJointeOrdre: number;         
    dossierPieceJointeDescription : string;  
    dossierPieceJointeDescriptionArabe : string;
    dossierPieceJointeEtat: number;         
  }
  